+++ src/components/index.jsx
@@ -21,13 +21,10 @@ export { default as SectionTitle } from './domain/SectionTitle';
 export { default as ArticleList } from './domain/ArticleList';
 export { default as CommentList } from './domain/CommentList';
 export { default as Radio } from './commons/Radio/ButtonRadio';
-export { default as SeriesEditor } from './domain/SeriesEditor';
 export { default as ConfirmCancleButtons } from './domain/ConfirmCancleButtons';
 export { default as CardSlider } from './domain/CardSlider';
 export { default as ImageUpload } from './domain/ImageUpload';
 export { default as Title } from './commons/Title';
-export { default as Period } from './domain/Period';
-export { default as ArticleEditor } from './domain/ArticleEditor';
 export { default as IconWrapper } from './commons/IconWrapper';
 export { default as LikeToggle } from './commons/LikeToggle';
 export { default as DetailForm } from './domain/DetailForm';
@@ -41,3 +38,4 @@ export { default as Loading } from './commons/Loading';
 export { default as NoData } from './commons/NoData';
 export { default as FollowListItem } from './domain/FollowListItem';
 export { default as SeriesForm } from './domain/SeriesForm';
+export { default as ArticleForm } from './domain/ArticleForm';
